+++ Source/ProjectFirma.Web/ProjectFirma.Web.csproj
@@ -1066,11 +1066,11 @@
     <Compile Include="Views\ProjectImage\New.cs" />
     <Compile Include="Views\ProjectImage\NewViewData.cs" />
     <Compile Include="Views\ProjectImage\NewViewModel.cs" />
-    <Compile Include="Views\ProjectOrganization\EditOrganizations.cs" />
-    <Compile Include="Views\ProjectOrganization\EditOrganizationsViewData.cs" />
-    <Compile Include="Views\ProjectOrganization\EditOrganizationsViewModel.cs" />
-    <Compile Include="Views\ProjectOrganization\EditOrganizationsViewModelTest.cs" />
-    <Compile Include="Views\ProjectOrganization\EditOrganizationsViewModelValidatorTest.cs" />
+    <Compile Include="Views\Shared\ProjectOrganization\EditOrganizations.cs" />
+    <Compile Include="Views\Shared\ProjectOrganization\EditOrganizationsViewData.cs" />
+    <Compile Include="Views\Shared\ProjectOrganization\EditOrganizationsViewModel.cs" />
+    <Compile Include="Views\Shared\ProjectOrganization\EditOrganizationsViewModelTest.cs" />
+    <Compile Include="Views\Shared\ProjectOrganization\EditOrganizationsViewModelValidatorTest.cs" />
     <Compile Include="Views\ProjectUpdate\Basics.cs" />
     <Compile Include="Views\ProjectUpdate\BasicsValidationResult.cs" />
     <Compile Include="Views\ProjectUpdate\BasicsValidationResultsTest.cs" />
@@ -1366,7 +1366,7 @@
     <Content Include="Views\ProjectCreate\ExpendituresController.js" />
     <Content Include="Views\ProjectCreate\PerformanceMeasuresController.js" />
     <Content Include="Views\ProjectFundingSourceRequest\ProjectFundingSourceRequestController.js" />
-    <Content Include="Views\ProjectOrganization\ProjectOrganizationController.js" />
+    <Content Include="Views\Shared\ProjectOrganization\ProjectOrganizationController.js" />
     <Content Include="Views\ProjectUpdate\ExpectedFundingController.js" />
     <Content Include="Views\Shared\ProjectLocationControls\ProjectLocationSimpleController.js" />
     <Content Include="Views\Shared\ProjectWatershedControls\EditProjectWatershedsController.js" />
@@ -1937,7 +1937,7 @@
     <Content Include="Views\ProjectBudget\ProjectBudgetController.js" />
     <Content Include="Views\ProjectExternalLink\ProjectExternalLinkController.js" />
     <Content Include="Views\ProjectFundingSourceExpenditure\ProjectFundingSourceExpenditureController.js" />
-    <Content Include="Views\ProjectOrganization\EditOrganizationsViewModelTest.AllViewModelFieldsAreSetFromConstructorTest.approved.txt" />
+    <Content Include="Views\Shared\ProjectOrganization\EditOrganizationsViewModelTest.AllViewModelFieldsAreSetFromConstructorTest.approved.txt" />
     <Content Include="Views\ProjectUpdate\PerformanceMeasuresController.js" />
     <Content Include="Views\ProjectUpdate\ExpendituresController.js" />
     <Content Include="Views\ProjectUpdate\BudgetsController.js" />
@@ -3226,7 +3226,7 @@
     <Content Include="Views\ProjectImageUpdate\New.cshtml" />
     <Content Include="Views\ProjectImage\Edit.cshtml" />
     <Content Include="Views\ProjectImage\New.cshtml" />
-    <Content Include="Views\ProjectOrganization\EditOrganizations.cshtml" />
+    <Content Include="Views\Shared\ProjectOrganization\EditOrganizations.cshtml" />
     <Content Include="Views\ProjectUpdate\Basics.cshtml" />
     <Content Include="Views\ProjectUpdate\CustomNotification.cshtml" />
     <Content Include="Views\ProjectUpdate\PerformanceMeasures.cshtml" />
