+++ Source/ProjectFirma.Web/Views/ProjectUpdate/Organizations.cshtml
@@ -2,11 +2,45 @@
 @using LtInfo.Common.ModalDialog
 @using ProjectFirma.Web.Models
 @using ProjectFirma.Web.Views.ProjectUpdate
+@using ProjectFirma.Web.Views.Shared.ProjectOrganization
 @inherits Organizations
 
 
 @if (ViewDataTyped.IsEditable)
 {
+@section JavascriptAndStylesContent
+{
+    <script type="text/javascript" src="@Url.Content("~/Views/Shared/ProjectOrganization/ProjectOrganizationController.js")"></script>
+    <script type="text/javascript" src="@Url.Content("~/Content/Bootstrap-select/bootstrap-select.min.js")"></script>
+    <script type="text/javascript">
+        // <![CDATA[
+        jQuery(document).ready(function() {
+            HookupCheckIfFormIsDirty(undefined, ".submitProposal");
+        });
+
+        function saveForm(autoAdvance) {
+            var form = jQuery("form");
+            // Do not submit if the form
+            // does not pass client side validation
+
+            if (!form.valid()) {
+                return false;
+            }
+            jQuery("<input type='hidden' name='@Html.NameFor(x => x.AutoAdvance)' value=" + autoAdvance + " />")
+                .appendTo(form);
+            form.submit();
+            return true;
+        }
+        // ]]>
+    </script>
+}
+    <div class="row">
+        <div class="col-xs-12">
+            @{ Html.RenderRazorSitkaPartial<EditOrganizations, EditOrganizationsViewData, EditOrganizationsViewModel>(ViewDataTyped.EditOrganizationsViewData, Model); }
+
+        </div>
+    </div>
+
     if (ViewDataTyped.ShowApproveAndReturnButton)
     {
         <hr/>
