+++ Source/ProjectFirma.Web/Views/ProjectUpdate/OrganizationsViewModel.cs
@@ -4,16 +4,15 @@ using System.ComponentModel.DataAnnotations;
 using System.Linq;
 using LtInfo.Common.Models;
 using ProjectFirma.Web.Models;
+using ProjectFirma.Web.Views.Shared.ProjectOrganization;
 
 namespace ProjectFirma.Web.Views.ProjectUpdate
 {
-    public class OrganizationsViewModel : FormViewModel
+    public class OrganizationsViewModel : EditOrganizationsViewModel
     {
         [DisplayName("Comments")]
         [StringLength(ProjectUpdateBatch.FieldLengths.ExpendituresComment)]
         public string Comments { get; set; }
-
-        public List<ProjectOrganizationSimple> ProjectOrganizationSimples { get; set; }
 
         /// <summary>
         /// Required by the ModelBinder
